feat: update version to 0.1.0 and refactor imports to use @buntal/http and @buntal/middlewares
